Reorganize tests folders/files to intentionally sort PR diffs
This puts `*.feature` files first, to review what the application should
do. Then the Behat Context files, to review that the test code does what
is says. Then the application code later, to review the application's
implementation details after confirming that the tests are correct.
