+++ application/_tests/bootstrap/NotificationContext.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sil\Idp\IdSync\Behat\Context;
+namespace Sil\Idp\IdSync\_tests\bootstrap;
 
 use Exception;
 use PHPUnit\Framework\Assert;
