+++ application/_tests/bootstrap/SecureUserIntegrationContext.php
@@ -1,11 +1,9 @@
 <?php
 
-namespace Sil\Idp\IdSync\Behat\Context;
+namespace Sil\Idp\IdSync\_tests\bootstrap;
 
 use Sil\Idp\IdSync\common\components\adapters\SecureUserIdStore;
-use Sil\Idp\IdSync\common\models\User;
 use Sil\PhpEnv\Env;
-use yii\base\BaseObject;
 
 /**
  * Defines application features from the specific context.
