+++ application/common/config/bootstrap.php
@@ -3,4 +3,4 @@
 Yii::setAlias('common', dirname(__DIR__));
 Yii::setAlias('backend', dirname(dirname(__DIR__)) . '/backend');
 Yii::setAlias('console', dirname(dirname(__DIR__)) . '/console');
-Yii::setAlias('@tests', dirname(dirname(__DIR__)) . '/tests');
+Yii::setAlias('@tests', dirname(dirname(__DIR__)) . '/_tests');
