--- application/features/behat.yml
@@ -1,34 +1,0 @@
-default:
-    suites:
-        google_sheets_integration_features:
-            paths:    [ "%paths.base%/../features/id-store-integration.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\GoogleSheetsIntegrationContext ]
-        idp_id_broker_integration_features:
-            paths:    [ "%paths.base%/../features/idp-id-broker-integration.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\IdpIdBrokerIntegrationContext ]
-        safety_cutoff_features:
-            paths:    [ "%paths.base%/../features/safetyCutoff.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\SafetyCutoffContext ]
-        sync_features:
-            paths:    [ "%paths.base%/../features/sync.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\SyncContext ]
-        user_features:
-            paths:    [ "%paths.base%/../features/user.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\UserContext ]
-        workday_integration_features:
-            filters:
-                # removing whitespace to silence faulty deprecation warning
-                # see https://github.com/Behat/Gherkin/pull/215#issuecomment-944115733
-                tags: '~@specificUser&&~@usersChangedSince&&~@canUpdateLastSynced'
-            paths:    [ "%paths.base%/../features/id-store-integration.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\WorkdayIntegrationContext ]
-        secure_user_integration_features:
-            filters:
-                # removing whitespace to silence faulty deprecation warning
-                # see https://github.com/Behat/Gherkin/pull/215#issuecomment-944115733
-                tags: '~@usersChangedSince&&~@canUpdateLastSynced'
-            paths:    [ "%paths.base%/../features/id-store-integration.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\SecureUserIntegrationContext ]
-        notification_features:
-            paths:    [ "%paths.base%/../features/notification.feature" ]
-            contexts: [ Sil\Idp\IdSync\Behat\Context\NotificationContext ]
